refactor(trade-in): decompose TradeInVoucherService into VoucherNumber/HtmlBuilder/PdfRenderer
Behavior-preserving (1 $tx, 0 JE). Facade keeps allocate/renderPdf + (prisma,storage)
ctor + delegates; internally constructs 3 sub-services (module + 3 consumers untouched).
allocate $tx + P2002 retry + idempotency moved WHOLE; ~700-LOC HTML/CSS template lifted
to VoucherHtmlBuilder VERBATIM; static sharedBrowser singleton kept on VoucherPdfRenderer;
voucherPrintedAt non-tx stamp ordering preserved. tsc 0, trade-in tests green, lint 0.
